Fixed test breaks
