+++ tests/unit/data/ar/Customer.php
@@ -1,6 +1,7 @@
 <?php
 namespace yiiunit\data\ar;
 
+use yii\db\ActiveQuery;
 use yiiunit\framework\db\ActiveRecordTest;
 
 /**
@@ -29,15 +30,15 @@ class Customer extends ActiveRecord
 		return $this->hasMany(Order::className(), ['customer_id' => 'id'])->orderBy('id');
 	}
 
-	public static function active($query)
-	{
-		$query->andWhere('status=1');
-	}
-
 	public function afterSave($insert)
 	{
 		ActiveRecordTest::$afterSaveInsert = $insert;
 		ActiveRecordTest::$afterSaveNewRecord = $this->isNewRecord;
 		parent::afterSave($insert);
+	}
+
+	public static function createQuery()
+	{
+		return new CustomerQuery(['modelClass' => get_called_class()]);
 	}
 }
