+++ tests/unit/data/ar/CustomerQuery.php
@@ -1,0 +1,15 @@
+<?php
+namespace yiiunit\data\ar;
+use yii\db\ActiveQuery;
+
+/**
+ * CustomerQuery
+ */
+class CustomerQuery extends ActiveQuery
+{
+	public static function active($query)
+	{
+		$query->andWhere('status=1');
+	}
+}
+
