+++ tests/unit/data/ar/mongodb/Customer.php
@@ -2,6 +2,8 @@
 
 namespace yiiunit\data\ar\mongodb;
 
+use yii\mongodb\ActiveQuery;
+
 class Customer extends ActiveRecord
 {
 	public static function collectionName()
@@ -20,13 +22,13 @@ class Customer extends ActiveRecord
 		];
 	}
 
-	public static function activeOnly($query)
-	{
-		$query->andWhere(['status' => 2]);
-	}
-
 	public function getOrders()
 	{
 		return $this->hasMany(CustomerOrder::className(), ['customer_id' => '_id']);
+	}
+
+	public static function createQuery()
+	{
+		return new CustomerQuery(['modelClass' => get_called_class()]);
 	}
 }
