+++ tests/unit/data/ar/sphinx/ArticleIndexQuery.php
@@ -1,0 +1,15 @@
+<?php
+namespace yiiunit\data\ar\sphinx;
+use yii\sphinx\ActiveQuery;
+
+/**
+ * ArticleIndexQuery
+ */
+class ArticleIndexQuery extends ActiveQuery
+{
+	public static function favoriteAuthor($query)
+	{
+		$query->andWhere('author_id=1');
+	}
+}
+
